+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Version 0.0.1
 
-- To be completed.
+- Updated `gillespie_model.jl` to contain and plot results of a basic stochastic
+  simulation of mtDNA population dynamics.